+++ package.json
@@ -2,7 +2,7 @@
   "name": "caesar",
   "version": "1.0.0",
   "description": "RS School NodeJS course",
-  "main": "index.js",
+  "main": "index.mjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
